change read() return type from uint8_t to int

+++ simple_gpio_class.h
@@ -6,6 +6,6 @@ struct gpio
 {
     gpio(uint8_t pin_arg, uint8_t mode) : pin(pin_arg) { pinMode(pin, mode); };
     void write(uint8_t value) const { digitalWrite(pin, value); };
-    uint8_t read() const { return(digitalRead(pin)); };
+    int read() const { return(digitalRead(pin)); };
     const uint8_t pin;
 };
